+++ src/lib/MuncherSettings.js
@@ -216,6 +216,38 @@ const MuncherSettings = {
 
     const advancedImportConfig = [
       {
+        name: "import-full-spell-list",
+        isChecked: game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-import-full-spell-list"),
+        title: "Import full spell list",
+        description:
+          "If a Cobalt Cookie is set, import all available spells for classes such as Cleric, where spells can be changed out on a long rest.",
+        enabled: true,
+      },
+      {
+        name: "use-active-sources",
+        isChecked: game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-use-active-sources"),
+        title: "Restrict imported spells to allowed sources",
+        description:
+          "Use allowed source on DDB when adding spells to character.",
+        enabled: true,
+      },
+      {
+        name: "remove-2024",
+        isChecked: game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-remove-2024"),
+        title: "Remove 2024 Spells",
+        description:
+          "Removes 2024 spells from character imports, and removes (Legacy) suffix on spell names.",
+        enabled: true,
+      },
+      {
+        name: "remove-legacy",
+        isChecked: game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-remove-legacy"),
+        title: "Remove Legacy Spells",
+        description:
+          "Removes legacy spells from character imports.",
+        enabled: true,
+      },
+      {
         name: "use-override",
         isChecked: game.settings.get(SETTINGS.MODULE_ID, "character-update-policy-use-override"),
         title: "Replace Items using those in your Override compendium",
@@ -400,6 +432,14 @@ const MuncherSettings = {
     if (selection === "use-combined-description") {
       game.settings.set(SETTINGS.MODULE_ID, "add-description-to-chat", false);
       MuncherSettings.toggleByName("character-import-policy-add-description-to-chat", false);
+    }
+
+    if (selection === "remove-2024" && checked) {
+      game.settings.set(SETTINGS.MODULE_ID, "character-update-policy-remove-legacy", false);
+      MuncherSettings.toggleByName("character-import-policy-remove-legacy", false);
+    } else if (selection === "remove-legacy" && checked) {
+      game.settings.set(SETTINGS.MODULE_ID, "character-update-policy-remove-2024", false);
+      MuncherSettings.toggleByName("character-import-policy-remove-2024", false);
     }
   },